Added component for IRB Board section.
